+++ src/state/text.rs
@@ -356,7 +356,7 @@ impl State {
         let plus = *self.code::<bool>();
         let note = *self.code::<bool>();
         let dir = *self.code::<i8>();
-        let width = *self.get_stack::<i64>() as i32;
+        let width = *self.get_stack::<i64>();
         let val = *self.get_stack::<i64>();
         let s = self.string_mut(pos - 16);
         ops::format_long(s, val, radix, width, token, plus, note, dir);
@@ -369,7 +369,7 @@ impl State {
         let plus = *self.code::<bool>();
         let note = *self.code::<bool>();
         let dir = *self.code::<i8>();
-        let width = *self.get_stack::<i64>() as i32;
+        let width = *self.get_stack::<i64>();
         let val = *self.get_stack::<i64>();
         let s = self.string_ref_mut(pos - 16);
         ops::format_long(s, val, radix, width, token, plus, note, dir);
@@ -378,8 +378,8 @@ impl State {
     pub fn format_float(&mut self) {
         let pos = *self.code::<u16>();
         let dir = *self.code::<i8>();
-        let precision = *self.get_stack::<i64>() as i32;
-        let width = *self.get_stack::<i64>() as i32;
+        let precision = *self.get_stack::<i64>();
+        let width = *self.get_stack::<i64>();
         let val = *self.get_stack::<f64>();
         let s = self.string_mut(pos - 24);
         ops::format_float(s, val, width, precision, dir);
@@ -388,8 +388,8 @@ impl State {
     pub fn format_stack_float(&mut self) {
         let pos = *self.code::<u16>();
         let dir = *self.code::<i8>();
-        let precision = *self.get_stack::<i64>() as i32;
-        let width = *self.get_stack::<i64>() as i32;
+        let precision = *self.get_stack::<i64>();
+        let width = *self.get_stack::<i64>();
         let val = *self.get_stack::<f64>();
         let s = self.string_ref_mut(pos - 24); // f64(8)+i64(8)+i64(8) = 24 bytes popped
         ops::format_float(s, val, width, precision, dir);
@@ -398,8 +398,8 @@ impl State {
     pub fn format_single(&mut self) {
         let pos = *self.code::<u16>();
         let dir = *self.code::<i8>();
-        let precision = *self.get_stack::<i64>() as i32;
-        let width = *self.get_stack::<i64>() as i32;
+        let precision = *self.get_stack::<i64>();
+        let width = *self.get_stack::<i64>();
         let val = *self.get_stack::<f32>();
         let s = self.string_mut(pos - 20);
         ops::format_single(s, val, width, precision, dir);
@@ -408,8 +408,8 @@ impl State {
     pub fn format_stack_single(&mut self) {
         let pos = *self.code::<u16>();
         let dir = *self.code::<i8>();
-        let precision = *self.get_stack::<i64>() as i32;
-        let width = *self.get_stack::<i64>() as i32;
+        let precision = *self.get_stack::<i64>();
+        let width = *self.get_stack::<i64>();
         let val = *self.get_stack::<f32>();
         let s = self.string_ref_mut(pos - 20);
         ops::format_single(s, val, width, precision, dir);
@@ -422,7 +422,7 @@ impl State {
         let width = *self.get_stack::<i64>();
         let val = self.string();
         let s = self.string_mut(pos - 8 - size_ptr() as u16);
-        ops::format_text(s, val.str(), width as i32, dir, token);
+        ops::format_text(s, val.str(), width, dir, token);
     }
 
     pub fn format_stack_text(&mut self) {
@@ -432,6 +432,6 @@ impl State {
         let width = *self.get_stack::<i64>();
         let val = self.string();
         let s = self.string_ref_mut(pos - 8 - size_ptr() as u16);
-        ops::format_text(s, val.str(), width as i32, dir, token);
+        ops::format_text(s, val.str(), width, dir, token);
     }
 }
